Update patch set 1
Patch Set 1:

This seems fine to me, however the blueprints are all going to change soon as a result of Kubernetes. For example the node SelectedDockerhost is going away. So this blueprint could be deprecated or at the very least this change could have been part of the k8 rewrite. But, no complaints.

Patch-set: 1