Update patch set 1
Patch Set 1:

renaming and/or changing the blueprints for k8s is outside of the scope of this change.  
That should be done in accord with the code that consumes the blueprint yaml files as a separate user story when it is ready.  

This change is only to have the latest versions of deployment-handler + policy-handler plus the new config for policy-handler catch_up logic

Patch-set: 1