Update patch set 7
Patch Set 7:

> (1 comment)
 > 
 > Generaly I would recommend to try this on lab to avoid this kind of
 > problems.

Hi Kornel,

We have validated this blueprint on the dcae-cloudify-manager pod in our own ONAP environment. Also we don't think this container_port parameter is mandatory as k8s-ves.yaml-template also uses the node type ContainerizedPlatformComponent but doesn't use container_port. If this blueprint doesn't work in your lab environment we are happy to change it, but we are not sure on how to add two ports with the container_port parameter as this is required.

Patch-set: 7
Label: Code-Review=0