Update patch set 7
Patch Set 7:

> Patch Set 3: Code-Review-1
> 
> (7 comments)
> 
> Include bootstrap version revision with this change.
> 
> For components not relying on dynamic topics, pls use dcae.nodes.ContainerizedServiceComponent nodetype in the blueprints

I don't see any change yet to the bootstrap container version. The pom.xml and version.properties files would need to be changed.

We build all of the blueprints into the bootstrap container Docker image.  So when we update any blueprint, we bump the version on the container, so that we don't end up with two different images with the same version tag.  The latest version of the container is 2.1.0, which is in the process of being released, so the release should be incremented to 2.1.1.

Patch-set: 7
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>