Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I think 
> external_cert_cert_type
> would be better named as
> external_cert_type
> but that can be dealt with separately.
> 
> I won't make this a blocking comment

I agree, but we want to keep inputs in line with current bp-gen. There were many changes of naming and input parameters etc during implementation of that and it ended up with a not very good sounding name. We'll probably try to address that some time in the future.

Patch-set: 1
Reviewer: Gerrit User 6044 <6044@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>