Update patch set 1
Patch Set 1:

Once we have merged a set of changes under a particular version (in this case 3.0.2), we should make any additional changes under a new version (in this case, it would have been 3.0.3).  We could have decided not to release 3.0.2 if it wasn't needed for anything.  It's especially important to be careful about this when more than one person is making changes.  

I will find the later snapshot and update the OOM WIP to use the latest snapshot so that we can make sure it goes through gating.   If it passes, I'll submit a new release.  If it doesn't pass, whoever made the last set of changes will need to figure out what went wrong.

Patch-set: 1